requirements.txt
playwright==1.44.0
pandas==2.2.2
python-decouple==3.8
selectolax==0.3.30
google-generativeai==0.7.1
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
multidict==6.0.5
psutil==5.9.8
pyee==11.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
python-dotenv==1.0.1
pytest
pytest-asyncio